This reduces allocations by using a `Stack` instead of a `List` for [ChannelExecutor](https://github.com/zio/zio/blob/cad5cafe3448f784092f354713f8b9fe4cb0ed51/streams/shared/src/main/scala/zio/stream/internal/ChannelExecutor.scala)'s `doneStack`. I measured the difference by running `ZStream.unit.forever.runDrain` and profiling the same way I did at https://github.com/zio/zio/pull/7365.

### benchmarks

before

```
[info] Benchmark                                   (chunkCount)  (chunkSize)   Mode  Cnt  Score    Error  Units
[info] StreamBenchmarks.zioChunkChunkFilterMapSum         10000         5000  thrpt   15  3.299 ±  0.026  ops/s
[info] StreamBenchmarks.zioChunkFilterMapSum              10000         5000  thrpt   15  2.466 ±  0.055  ops/s
[info] StreamBenchmarks.zioGroupAdjecentBy                10000         5000  thrpt   15  1.083 ±  0.003  ops/s
[info] StreamBenchmarks.zioGroupByKey                     10000         5000  thrpt   15  2.533 ±  0.021  ops/s
[info] StreamBenchmarks.zioGroupWithin                    10000         5000  thrpt   15  0.033 ±  0.001  ops/s
[info] StreamBenchmarks.zioMapAccum                       10000         5000  thrpt   15  2.227 ±  0.020  ops/s
[info] StreamBenchmarks.zioSliding                        10000         5000  thrpt   15  0.117 ±  0.001  ops/s
[info] StreamBenchmarks.zioTakeWhile                      10000         5000  thrpt   15  3.212 ±  0.013  ops/s
```

after
```
[info] Benchmark                                   (chunkCount)  (chunkSize)   Mode  Cnt  Score   Error  Units
[info] StreamBenchmarks.zioChunkChunkFilterMapSum         10000         5000  thrpt   15  3.323 ± 0.042  ops/s
[info] StreamBenchmarks.zioChunkFilterMapSum              10000         5000  thrpt   15  2.512 ± 0.012  ops/s
[info] StreamBenchmarks.zioGroupAdjecentBy                10000         5000  thrpt   15  1.073 ± 0.003  ops/s
[info] StreamBenchmarks.zioGroupByKey                     10000         5000  thrpt   15  2.453 ± 0.187  ops/s
[info] StreamBenchmarks.zioGroupWithin                    10000         5000  thrpt   15  0.032 ± 0.001  ops/s
[info] StreamBenchmarks.zioMapAccum                       10000         5000  thrpt   15  2.194 ± 0.012  ops/s
[info] StreamBenchmarks.zioSliding                        10000         5000  thrpt   15  0.108 ± 0.001  ops/s
[info] StreamBenchmarks.zioTakeWhile                      10000         5000  thrpt   15  3.748 ± 0.025  ops/s
```

### allocations

before

![forever before](https://user-images.githubusercontent.com/1625822/204153455-f9eb9d28-1381-4f87-8bbf-ca7683675d11.png)

after

Notice that the `$colon$colon` allocations disappear, and nothing takes their place.

![forever after](https://user-images.githubusercontent.com/1625822/204153458-a747e5c6-5257-4e4f-8fb3-5b1efd92a59f.png)